performance improvements, ruff fixes

+++ src/qilisdk_cpp/backends/qilisim/config/qilisim_config.h
@@ -36,7 +36,7 @@ class QiliSimConfig {
     double adaptive_tol = 1e-2;
     bool measurement_collapse = true;
     int max_terms = 1000;
-    float order = 2;
+    int order = 2;
     int shots = 1000;
     int warmups = 10;
 
@@ -58,7 +58,7 @@ class QiliSimConfig {
     bool get_combine_single_qubit_gates() const { return combine_single_qubit_gates; }
     bool get_measurement_collapse() const { return measurement_collapse; }
     int get_max_terms() const { return max_terms; }
-    float get_order() const { return order; }
+    int get_order() const { return order; }
     int get_shots() const { return shots; }
     int get_warmups() const { return warmups; }
 
@@ -79,7 +79,7 @@ class QiliSimConfig {
     void set_combine_single_qubit_gates(bool value) { combine_single_qubit_gates = value; }
     void set_measurement_collapse(bool value) { measurement_collapse = value; }
     void set_max_terms(int value) { max_terms = value; }
-    void set_order(float value) { order = value; }
+    void set_order(int value) { order = value; }
     void set_shots(int value) { shots = value; }
     void set_warmups(int value) { warmups = value; }
